move schematic layer interfaces file
